// src/lib/proto/base/base.ts
export enum AccountType {
  /** ACCOUNT_TYPE_UNSPECIFIED - Default */
  ACCOUNT_TYPE_UNSPECIFIED = 0,
  ACCOUNT_TYPE_ASSET = 1,
  ACCOUNT_TYPE_LIABILITY = 2,
  ACCOUNT_TYPE_INCOME = 3,
  ACCOUNT_TYPE_EXPENSE = 4,
  ACCOUNT_TYPE_EQUITY = 5,
  UNRECOGNIZED = -1,
}
// ...
export enum TransactionType {
  /** TRANSACTION_TYPE_UNSPECIFIED - Default */
  TRANSACTION_TYPE_UNSPECIFIED = 0,
  TRANSACTION_TYPE_INCOME = 1,
  TRANSACTION_TYPE_EXPENSE = 2,
  TRANSACTION_TYPE_TRANSFER = 3,
  TRANSACTION_TYPE_OPENING_BALANCE = 4,
  UNRECOGNIZED = -1,
}